ai_agents_service/requirements.txt
langgraph>=0.0.25
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-ollama>=1.0.1
duckduckgo-search>=4.1.1
arxiv>=2.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.0
docker>=7.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
